Move middleware to live on Raven prototype
