+++ lib/client.js
@@ -359,6 +359,32 @@ extend(Raven.prototype, {
    */
   setShouldSendCallback: function (callback) {
     return this.setCallbackHelper('shouldSendCallback', callback);
+  },
+
+  requestHandler: function () {
+    var self = this;
+    return function (req, res, next) {
+      self.context({}, next, next);
+    };
+  },
+
+  errorHandler: function () {
+    var self = this;
+    return function (err, req, res, next) {
+      var status = err.status || err.statusCode || err.status_code || 500;
+
+      // skip anything not marked as an internal server error
+      if (status < 500) return next(err);
+
+      var kwargs = parsers.parseRequest(req);
+      if (domain.active && domain.active.sentryContext) {
+        kwargs = extend(kwargs, domain.active.sentryContext);
+      }
+      return self.captureException(err, kwargs, function (sendErr, eventId) {
+        res.sentry = eventId;
+        next(err);
+      });
+    };
   }
 });
 
@@ -383,6 +409,7 @@ Raven.prototype.get_ident = Raven.prototype.getIdent;
 
 // Maintain old API compat, need to make sure arguments length is preserved
 function Client(dsn, options) {
+  if (dsn instanceof Client) return dsn;
   var ravenInstance = new Raven();
   return ravenInstance.config.apply(ravenInstance, arguments);
 }
