+++ lib/middleware/connect.js
@@ -1,9 +1,6 @@
 'use strict';
 
-var raven = require('../client');
-var parsers = require('../parsers');
-var extend = require('../utils').extend;
-var domain = require('domain');
+var Raven = require('../client');
 
 // Legacy support
 var connectMiddleware = function (client) {
@@ -13,30 +10,15 @@ var connectMiddleware = function (client) {
 // Error handler. This should be the last item listed in middleware, but
 // before any other error handlers.
 connectMiddleware.errorHandler = function (client) {
-  // client = client instanceof raven.Client ? client : new raven.Client(client);
-  return function (err, req, res, next) {
-    var status = err.status || err.statusCode || err.status_code || 500;
-
-    // skip anything not marked as an internal server error
-    if (status < 500) return next(err);
-
-    var kwargs = parsers.parseRequest(req);
-    if (domain.active && domain.active.sentryContext) {
-      kwargs = extend(kwargs, domain.active.sentryContext);
-    }
-    return client.captureException(err, kwargs, function (err, eventId) {
-      res.sentry = eventId;
-      next(err);
-    });
-  };
+  client = client instanceof Raven.Client ? client : new Raven.Client(client);
+  return client.errorHandler();
 };
 
 // Ensures asynchronous exceptions are routed to the errorHandler. This
 // should be the **first** item listed in middleware.
 connectMiddleware.requestHandler = function (client) {
-  return function (req, res, next) {
-    client.context({}, next, next);
-  };
+  client = client instanceof Raven.Client ? client : new Raven.Client(client);
+  return client.requestHandler();
 };
 
 module.exports = connectMiddleware;
